requirements.txt
aiobotocore==2.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioitertools==0.12.0
aiosignal==1.4.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
boto3==1.38.4
botocore==1.38.25
certifi==2025.7.9
charset-normalizer==3.4.2
comm==0.2.2
DateTime==5.5
debugpy==1.8.14
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
frozenlist==1.7.0
fsspec==2025.5.1
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kafka-python==2.2.15
matplotlib-inline==0.1.7
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.0
requests==2.32.4
s3fs==2025.5.1
s3transfer==0.12.0
schedule==1.2.2
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.1
zope.interface==7.2
